split command package code
