+++ pkg/kt/command/mesh.go
@@ -1,0 +1,30 @@
+package command
+
+import (
+	"github.com/alibaba/kt-connect/pkg/kt/options"
+	"github.com/rs/zerolog"
+	"github.com/urfave/cli"
+)
+
+// newMeshCommand return new mesh command
+func newMeshCommand(options *options.DaemonOptions) cli.Command {
+	return cli.Command{
+		Name:  "mesh",
+		Usage: "mesh kubernetes deployment to local",
+		Flags: []cli.Flag{
+			cli.StringFlag{
+				Name:        "expose",
+				Usage:       "expose port",
+				Destination: &options.MeshOptions.Expose,
+			},
+		},
+		Action: func(c *cli.Context) error {
+			if options.Debug {
+				zerolog.SetGlobalLevel(zerolog.DebugLevel)
+			}
+			action := Action{}
+			action.Mesh(c.Args().First(), options)
+			return nil
+		},
+	}
+}
